backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Environment
python-dotenv==1.0.1

# Database
prisma==0.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Upload
python-multipart==0.0.9

# Data Validation
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

# Graph Algorithms
networkx==3.3

# NLP
nltk==3.8.1

# Async / Utilities
anyio==4.3.0
aiofiles==23.2.1
httpx==0.27.0
loguru==0.7.2
python-slugify==8.0.4